Simulation/requirements.txt
asgiref==3.2.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==2.0.7
django-cors-headers==3.1.1
django-filter==2.2.0
django-filters==0.2.1
django-redis==4.10.0
djangorestframework==3.11.0
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.2.0
psycopg2==2.8.5
PyMySQL==0.9.3
pytz==2020.1
redis==3.5.3
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
Werkzeug==1.0.1
xlrd==1.2.0
